- refactor section tab bar and split TabBarBlocBuilder
