+++ lib/features/home/ui/screens/home_screen.dart
@@ -1,16 +1,42 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:netflix_clone/core/helpers/spacing.dart';
+import 'package:netflix_clone/core/models/genre_model.dart';
 import 'package:netflix_clone/core/theming/app_colors.dart';
-import 'package:netflix_clone/features/home/logic/cubit/home_cubit.dart';
-import 'package:netflix_clone/features/home/logic/cubit/home_state.dart';
 import 'package:netflix_clone/features/home/ui/widgets/slider_section/slider_and_indicator_bloc_builder.dart';
-import 'package:netflix_clone/features/home/ui/widgets/tab_bar_section.dart';
+import 'package:netflix_clone/features/home/ui/widgets/tab_bar_widget.dart';
 import '../../../../core/helpers/constants.dart';
+import '../widgets/tab_bar_bloc_builder.dart';
 
-class HomeScreen extends StatelessWidget {
+class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
+
+  @override
+  State<HomeScreen> createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen>  with SingleTickerProviderStateMixin {
+
+    late TabController _tabController;
+ @override
+  void initState() {
+    super.initState();
+    _tabController = TabController(length: genresList.length, initialIndex: 0, vsync: this);
+    _tabController.addListener(_handleTabIndex);
+  }
+
+  @override
+  void dispose() {
+     _tabController.removeListener(_handleTabIndex);
+    _tabController.dispose();
+    super.dispose();
+  }
+
+  
+  void _handleTabIndex() {
+    setState(() {});
+  }
+
 
   @override
   Widget build(BuildContext context) {
@@ -42,21 +68,11 @@ class HomeScreen extends StatelessWidget {
           child: Column(
             children: [
               const SliderAndIndicatorBlocBuilder(),
-              BlocBuilder<HomeCubit, HomeState>(
-                builder: (context, state) {
-                  if (state is MoviesListLoading) {
-                    return const Center(
-                      child: CircularProgressIndicator(),
-                    );
-                  } else if (state is MoviesListSuccess) {
-                    return TabBarSection(moviesList: state.movies);
-                  } else if (state is MoviesListFailure) {
-                    return Text(state.error);
-                    
-                  }
-                  return const SizedBox.shrink();
-                },
-              ),
+
+              TabBarWidget(tabController: _tabController),
+
+              TabBarBlocBuilder(tabController: _tabController),
+              
               Container(),
             ],
           ),
