+++ lib/features/home/ui/widgets/slider_section/slider_and_indicator_bloc_builder.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:netflix_clone/features/home/logic/cubit/home_cubit.dart';
 
 import '../../../logic/cubit/home_state.dart';
@@ -17,8 +18,11 @@ class SliderAndIndicatorBlocBuilder extends StatelessWidget {
           current is PopularMoviesFailure,
       builder: (context, state) {
         if (state is PopularMoviesLoading) {
-          return const Center(
-            child: CircularProgressIndicator(),
+          return  SizedBox(
+            height: 200.h,
+            child: const Center(
+              child:  CircularProgressIndicator(),
+            ),
           );
         } else if (state is PopularMoviesSuccess) {
           return SliderAndIndicatorSection(list: state.movies.sublist(0, 5));
